+++ analysis/filter/StructureModifier/src/main/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/NodeRemover.kt
@@ -42,10 +42,10 @@ class NodeRemover(private val project: Project) {
         return edges.toMutableList()
     }
 
-    private fun copyAttributeTypes(): MutableMap<String, MutableList<Map<String, AttributeType>>> {
-        val mergedAttributeTypes: MutableMap<String, MutableList<Map<String, AttributeType>>> = mutableMapOf()
+    private fun copyAttributeTypes(): MutableMap<String, MutableMap<String, AttributeType>> {
+        val mergedAttributeTypes: MutableMap<String, MutableMap<String, AttributeType>> = mutableMapOf()
         project.attributeTypes.forEach {
-            mergedAttributeTypes[it.key] = it.value.toMutableList()
+            mergedAttributeTypes[it.key] = it.value
         }
         return mergedAttributeTypes.toMutableMap()
     }