+++ analysis/import/SCMLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/scmlogparser/input/metrics/AgeInWeeks.kt
@@ -1,6 +1,7 @@
 package de.maibornwolff.codecharta.importer.scmlogparser.input.metrics
 
 import de.maibornwolff.codecharta.importer.scmlogparser.input.Commit
+import de.maibornwolff.codecharta.model.AttributeType
 import java.time.OffsetDateTime
 
 class AgeInWeeks : Metric {
@@ -21,6 +22,10 @@ class AgeInWeeks : Metric {
     override fun value(): Number {
         val thisWeek = CalendarWeek.forDateTime(OffsetDateTime.now())
         return CalendarWeek.numberOfWeeksBetween(firstCommit, thisWeek)
+    }
+
+    override fun attributeType(): AttributeType {
+        return AttributeType.relative
     }
 
 }