+++ analysis/import/TokeiImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/tokeiimporter/TokeiImporter.kt
@@ -3,9 +3,7 @@ package de.maibornwolff.codecharta.importer.tokeiimporter
 import com.google.gson.Gson
 import com.google.gson.JsonElement
 import com.google.gson.JsonParser
-import de.maibornwolff.codecharta.model.MutableNode
-import de.maibornwolff.codecharta.model.PathFactory
-import de.maibornwolff.codecharta.model.ProjectBuilder
+import de.maibornwolff.codecharta.model.*
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.serialization.mapLines
 import mu.KotlinLogging
@@ -23,6 +21,12 @@ class TokeiImporter(private val input: InputStream = System.`in`,
                     private val error: PrintStream = System.err) : Callable<Void> {
     private var TOP_LEVEL_OBJECT: String = "inner"
     private val logger = KotlinLogging.logger {}
+
+    private val attributeTypes = AttributeTypes(type = "nodes")
+            .add("rloc", AttributeType.absolute)
+            .add("loc", AttributeType.absolute)
+            .add("empty_lines", AttributeType.absolute)
+            .add("comment_lines", AttributeType.absolute)
 
     private lateinit var projectBuilder: ProjectBuilder
 
@@ -57,6 +61,7 @@ class TokeiImporter(private val input: InputStream = System.`in`,
                 }
             }
         }
+        projectBuilder.addAttributeTypes(attributeTypes)
         ProjectSerializer.serializeProject(projectBuilder.build(), writer())
         return null
     }