+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/attributeTypes/AttributeTypes.kt
@@ -1,7 +1,0 @@
-package de.maibornwolff.codecharta.attributeTypes
-
-import de.maibornwolff.codecharta.model.AttributeType
-
-data class AttributeTypes(
-        val attributeTypes: Map<String, AttributeType>,
-        val type: String = "")