+++ visualization/app/codeCharta/state/edgeMetricData.service.ts
@@ -1,4 +1,4 @@
-import { MetricData, BlacklistItem, Edge, BlacklistType, CodeMapNode, EdgeMetricCount } from "../codeCharta.model"
+import { MetricData, BlacklistItem, Edge, BlacklistType, CodeMapNode, EdgeMetricCount, AttributeTypeValue } from "../codeCharta.model"
 import { IRootScopeService } from "angular"
 import { CodeMapHelper } from "../util/codeMapHelper"
 import { HierarchyNode } from "d3"
@@ -78,6 +78,10 @@ export class EdgeMetricDataService implements FilesSelectionSubscriber, Blacklis
 
 	public getMetricData(): MetricData[] {
 		return this.edgeMetricData
+	}
+
+	public getAttributeTypeByMetric(metricName: string): AttributeTypeValue {
+		return this.storeService.getState().fileSettings.attributeTypes.edges[metricName]
 	}
 
 	private calculateMetrics(): MetricData[] {