+++ visualization/app/codeCharta/state/store/fileSettings/attributeTypes/attributeTypes.merger.spec.ts
@@ -9,30 +9,30 @@ describe("AttributeTypesMerger", () => {
 
 		beforeEach(() => {
 			attributes1 = {
-				nodes: [
-					{ attribute1: AttributeTypeValue.relative },
-					{ attribute2: AttributeTypeValue.absolute },
-					{ attribute3: AttributeTypeValue.relative }
-				],
-				edges: [{ attribute4: AttributeTypeValue.relative }, { attribute5: AttributeTypeValue.absolute }]
+				nodes: {
+					attribute1: AttributeTypeValue.relative,
+					attribute2: AttributeTypeValue.absolute,
+					attribute3: AttributeTypeValue.relative
+				},
+				edges: { attribute4: AttributeTypeValue.relative, attribute5: AttributeTypeValue.absolute }
 			}
 
 			attributes2 = {
-				nodes: [
-					{ attribute1: AttributeTypeValue.absolute },
-					{ attribute3: AttributeTypeValue.relative },
-					{ attribute4: AttributeTypeValue.absolute }
-				],
-				edges: [
-					{ attribute4: AttributeTypeValue.absolute },
-					{ attribute6: AttributeTypeValue.absolute },
-					{ attribute7: AttributeTypeValue.relative }
-				]
+				nodes: {
+					attribute1: AttributeTypeValue.absolute,
+					attribute3: AttributeTypeValue.relative,
+					attribute4: AttributeTypeValue.absolute
+				},
+				edges: {
+					attribute4: AttributeTypeValue.absolute,
+					attribute6: AttributeTypeValue.absolute,
+					attribute7: AttributeTypeValue.relative
+				}
 			}
 
 			attributes3 = {
-				nodes: [{ attribute1: AttributeTypeValue.relative }, { attribute3: AttributeTypeValue.absolute }],
-				edges: []
+				nodes: { attribute1: AttributeTypeValue.relative, attribute3: AttributeTypeValue.absolute },
+				edges: {}
 			}
 		})
 
@@ -40,18 +40,18 @@ describe("AttributeTypesMerger", () => {
 			const attributeTypes = getMergedAttributeTypes([attributes1, attributes2])
 
 			const expected = {
-				nodes: [
-					{ attribute1: AttributeTypeValue.relative },
-					{ attribute2: AttributeTypeValue.absolute },
-					{ attribute3: AttributeTypeValue.relative },
-					{ attribute4: AttributeTypeValue.absolute }
-				],
-				edges: [
-					{ attribute4: AttributeTypeValue.relative },
-					{ attribute5: AttributeTypeValue.absolute },
-					{ attribute6: AttributeTypeValue.absolute },
-					{ attribute7: AttributeTypeValue.relative }
-				]
+				nodes: {
+					attribute1: AttributeTypeValue.relative,
+					attribute2: AttributeTypeValue.absolute,
+					attribute3: AttributeTypeValue.relative,
+					attribute4: AttributeTypeValue.absolute
+				},
+				edges: {
+					attribute4: AttributeTypeValue.relative,
+					attribute5: AttributeTypeValue.absolute,
+					attribute6: AttributeTypeValue.absolute,
+					attribute7: AttributeTypeValue.relative
+				}
 			}
 
 			expect(attributeTypes).toEqual(expected)
@@ -61,12 +61,12 @@ describe("AttributeTypesMerger", () => {
 			const attributeTypes = getMergedAttributeTypes([attributes1, attributes3])
 
 			const expected = {
-				nodes: [
-					{ attribute1: AttributeTypeValue.relative },
-					{ attribute2: AttributeTypeValue.absolute },
-					{ attribute3: AttributeTypeValue.relative }
-				],
-				edges: [{ attribute4: AttributeTypeValue.relative }, { attribute5: AttributeTypeValue.absolute }]
+				nodes: {
+					attribute1: AttributeTypeValue.relative,
+					attribute2: AttributeTypeValue.absolute,
+					attribute3: AttributeTypeValue.relative
+				},
+				edges: { attribute4: AttributeTypeValue.relative, attribute5: AttributeTypeValue.absolute }
 			}
 
 			expect(attributeTypes).toEqual(expected)