+++ CHANGELOG.md
@@ -9,6 +9,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added 🚀
 
+- Median symbol for aggregated relative metrics #365
+- AttributeTypes for tokeiImporter and SCMLogParser #365
+
+### Changed
+
+- Metrics with AttributeType relative are now aggregated using the median #365
+
+### Removed 🗑
+
+### Fixed 🐞
+
+- Consistency of AttributeTypes representation #365
+
+### Chore 👨‍💻 👩‍💻
+
+## [1.44.0] - 2020-03-27
+
+### Added 🚀
+
 - Dialog to select between different sorting options #388
 - Button to reverse the current selected sorting #388
 
@@ -17,6 +36,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Removed 🗑
 
 ### Fixed 🐞
+
+- Show file selection in toolBar after excluding or hiding a node instead of an empty toolBar #896
 
 ### Chore 👨‍💻 👩‍💻